lumi_project/requirements.txt
# AWS
boto3==1.34.0
botocore==1.34.0

# Environment
python-dotenv==1.0.0

# LangChain
langchain==0.2.0
langchain-aws==0.1.0
langchain-community==0.2.0

# Document generation (test docs)
Pillow==10.3.0

# Data & display
pandas==2.2.0
tabulate==0.9.0
tqdm==4.66.0
colorama==0.4.6
requests==2.31.0

# Phase 4 - RAG
faiss-cpu==1.8.0
tiktoken==0.7.0
